requirements.txt
streamlit
tensorflow>=2.0
pillow  # If you're using PIL or Image from PIL
